pyproject.toml
[project]
name = "business-chat-env"
version = "0.1.0"
description = "OpenEnv environment for business chat analysis"
authors = [
    {name = "Scaler Hackathon"}
]
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "uvicorn (>=0.43.0,<0.44.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "openai (>=2.30.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "requests (>=2.33.1,<3.0.0)",
    "orjson (>=3.11.8,<4.0.0)",
    "openenv-core (>=0.2.0)"
]

[project.scripts]
server = "server.app:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
